mixed routes
routes, sales controller
