+++ backend/controllers/salesController.js
@@ -1,63 +1,93 @@
 // TODO  all sales here
+// * NOT ADMIN Sales page
 
-// TODO all rentals here
+const { Op } = require('sequelize');
 
 const db = require('../config/config');
 
-// const {
-// 	findAllSalesById,
-// 	findAllRentalsById,
-// 	findRentalById,
-// } = require('../db/db');
-
 module.exports = {
 	// all products to sale
-	findAllProductsForSale: (req, res) => {
-		db.Products.findAll({
+	findAllProductsForSale: async (req, res) => {
+		await db.Products.findAll({
 			where: { units: { [Op.gt]: 0 } },
 		})
 			.then((results) => {
-				res.json(results);
+				res.status(200).json(results);
 				console.log(JSON.stringify(results, null, 2));
+				return results;
 			})
 			.catch((err) => console.error(err));
 	},
+
 	// all products to sale by category
-	findAllProductsByCategory: (req, res) => {
+	findAllProductsByCategory: async (req, res) => {
 		const category = req.params.category;
-		db.Products.findAll({
-			where: { category: { category } },
+		await db.Products.findAll({
+			where: {
+				[Op.and]: [{ category: category }, { units: { [Op.gt]: 0 } }],
+			},
 		})
 			.then((results) => {
-				res.json(results);
+				res.status(200).json(results);
 				console.log(JSON.stringify(results, null, 2));
 			})
 			.catch((err) => console.error(err));
 	},
+
 	// all products to sale by category
-	findAllProductsBySubCategory: (req, res) => {
-		const category = req.params.sub - category;
-		db.Products.findAll({
-			where: { sub_category: category },
+	findAllProductsBySubCategory: async (req, res) => {
+		const category = req.params.subcategory;
+		await db.Products.findAll({
+			where: {
+				[Op.and]: [{ sub_category: category }, { units: { [Op.gt]: 0 } }],
+			},
 		})
 			.then((results) => {
-				res.json(results);
+				res.status(200).json(results);
 				console.log(JSON.stringify(results, null, 2));
+				return results;
 			})
 			.catch((err) => console.error(err));
 	},
 
 	// TODO not correct
 	// all products to sale
-	findProductsById: (req, res) => {
+	findProductsById: async (req, res) => {
 		const id = req.params.id;
-		db.Products.findAll({
-			where: { id: id },
+		await db.Products.findAll({
+			where: {
+				[Op.and]: [{ id: id }, { units: { [Op.gt]: 0 } }],
+			},
 		})
 			.then((results) => {
-				res.json(results);
+				res.status(200).json(results);
 				console.log(JSON.stringify(results, null, 2));
+				return results;
 			})
 			.catch((err) => console.error(err));
 	},
+
+
+	// finds most recent 5 sales 
+	findRecentSales: async (req, res) => {
+		await db.Sales.findAll({
+			limit: 5,
+			order: [['createdAt', 'ASC']],
+			include: db.Products
+		})
+			.then((results) => res.status(200).json(results))
+			.catch((err) => console.error(err));
+	},
+
+	// find all sales with most recent first
+	findAllSales: async (req, res) => {
+		await db.Sales.findAll({
+			
+		//	order: [['createdAt', 'ASC']],
+		//	include: db.Products
+		})
+			.then((results) => res.status(200).json(results))
+			.catch((err) => console.error(err));
+	},
+
 };
