+++ backend/routes/sales.routes.js
@@ -4,8 +4,11 @@ const { salesController } = require('../controllers');
 
 router.get('/', salesController.findAllProductsForSale);
 router.get('/:id', salesController.findProductsById);
-router.get('/:category', salesController.findAllProductsByCategory);
-router.get('/:sub-category', salesController.findAllProductsBySubCategory);
-
-
+router.get('/category/:category', salesController.findAllProductsByCategory);
+router.get(
+	'/sub-category/:subcategory',
+	salesController.findAllProductsBySubCategory
+);
+router.get('/admin/all-sales', salesController.findAllSales); // all products for sale
+router.get('/admin/recent-sales', salesController.findRecentSales); // all products for sale
 module.exports = router;
